+++ lib/src/sources/secure_data_source.dart
@@ -1,6 +1,7 @@
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 import 'package:meta/meta.dart';
 import 'package:nosql_persistence/src/persistence_interface.dart';
+import 'package:nosql_persistence/src/resolvers/persistence_base_types_resolver.dart';
 import 'package:nosql_persistence/src/resolvers/persistence_expired_system_resolver.dart';
 import 'package:nosql_persistence/src/resolvers/persistence_migrations_resolver.dart';
 import 'package:nosql_persistence/src/resolvers/persistense_json_resolver.dart';
@@ -15,7 +16,8 @@ abstract base class SecureDataSource extends PersistenceInterface
     with
         PersistenseJsonResolver,
         PersistenceMigrationsResolver,
-        PersistenceExpiredSystemResolver {
+        PersistenceExpiredSystemResolver,
+        PersistenceBaseTypesResolver {
   final FlutterSecureStorage _secureStorage;
 
   @override