+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.3.0
+* Add getBool, putBool, getInt, putBool methods for new PersistenceBaseTypesResolver. It can be used in any datasources.
+* Add getJsonListTyped and putJsonListTyped on PersistenseJsonResolver.
+
 ## 0.2.0
 
 * Add the getExpired, putExpired, getExpiredTyped and putExpiredTyped methods for new expired system to the data sources